requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
Pillow==10.1.0
opencv-python==4.8.1.78
tensorflow==2.14.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
webdriver-manager==4.0.1
numpy==1.24.3
psutil==5.9.6
instaloader==4.10.3
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
scikit-image==0.21.0
nltk==3.8.1
textblob==0.17.1
imagehash==4.3.1
ultralytics==8.0.196
brisque==0.0.23
# Background Task Processing
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.4.0
